Add structure binding/unbinding
