+++ pons/__init__.py
@@ -1,5 +1,6 @@
 from .client import Client
-from .contract_abi import ContractABI, Method, MethodCall, uint256, Constructor
+from .contract_types import Struct, uint256
+from .contract_abi import ContractABI, Method, MethodCall, Constructor
 from .contract import CompiledContract, DeployedContract
 from .provider import HTTPProvider
 from .signer import Signer, AccountSigner
